Add files via upload
V3 is initial commit version. Versions 1 and 2 do not exist in the repository. File built as Live Script because this is intended to be used in a variety of datasets matching this format. The visibility allows newer matlab users to confidently use the script and perform small troubleshooting if error-correcting is not adequate or if data is not formatted correctly.